+++ src/components/Link/Link.stories.js
@@ -24,7 +24,7 @@ export default {
   },
   argTypes: {
     variant: {
-      options: ['link', 'primary-button', 'secondary-button'],
+      options: ['link', 'primary-button', 'secondary-button', 'subtle-button'],
       control: {
         type: 'select'
       }
@@ -64,4 +64,9 @@ StyledAsPrimaryButton.args = {
 export const StyledAsSecondaryButton = Template.bind({});
 StyledAsSecondaryButton.args = {
   variant: 'secondary-button'
+};
+
+export const StyledAsSubtleButton = Template.bind({});
+StyledAsSubtleButton.args = {
+  variant: 'subtle-button'
 };