+++ src/utils/index.js
@@ -2,3 +2,4 @@ export * from './array';
 export * from './date';
 export * from './keyboard';
 export * from './object';
+export * from './formatBytes';